Xamarin.Android guide
Documents #1.

The `..\..\..\..\..\bin\Release` path in `obfuscar.xml` is ugly. Ideally, obfuscar.exe would take parameters for the paths so that the MSBuild properties could be used. Or even better would be if obfuscar were distributed as a .NET Standard DLL that could be executed as a MSBuild task directly. Then it could be used to build on Mac as well.